+++ src/lib.rs
@@ -42,9 +42,10 @@ pub mod types {
     pub use super::emails::{
         Attachment, CreateEmailOptions, EmailEvent, EmailEventsData, EmailOptions, EmailState,
         EventType, GeoIp, GetEmailResponse, ListEmailEventsOptions, ListEmailEventsResponse,
-        ListEmailsOptions, ListEmailsResponse, Pagination, QuotaInfo, ScheduleEmailOptions,
-        ScheduledEmailState, ScheduledTransmission, SendEmailResponse, SendEmailWithQuotaResponse,
-        SentEmailEventsData, SentEmailListItem, UserAgentParsed,
+        ListEmailsOptions, ListEmailsResponse, ListScheduledEmailsOptions,
+        ListScheduledEmailsResponse, Pagination, QuotaInfo, ScheduleEmailOptions, ScheduledEmail,
+        ScheduledEmailPagination, ScheduledEmailState, ScheduledEmailWithQuota, SendEmailResponse,
+        SendEmailWithQuotaResponse, SentEmailEventsData, SentEmailListItem, UserAgentParsed,
     };
 
     // Domains